requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0.2
django-admin-interface==0.28.5
django-cleanup==8.0.0
django-colorfield==0.11.0
django-mathfilters==1.0.0
django-paypal==2.0
django-shortuuidfield==0.1.3
django-taggit==4.0.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
idna==3.6
Pillow==10.0.1
python-slugify==8.0.4
pytz==2023.3.post1
requests==2.31.0
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
tzdata==2023.3
urllib3==2.1.0